clean structure for PCA9685Driver test
